refactor: move fetch notification into NotificationPopover
