+++ src/modules/notification/components/NotificationPopover.tsx
@@ -1,24 +1,76 @@
-import React, {useState} from "react";
+import React, {useEffect, useState} from "react";
 import {Card, CardContent, CardHeader} from "@/components/ui/card.tsx";
 import {Bell, Check, X} from "lucide-react";
 import {Notification, NotificationStatus} from '@/core/types/notifications.ts';
 import {formatTimeAgo} from "@/core/utils/timeAgo.ts";
 import {Tooltip, TooltipContent, TooltipProvider, TooltipTrigger} from "@/components/ui/tooltip.tsx";
 import {Dialog, DialogContent, DialogHeader, DialogTitle} from "@/components/ui/dialog.tsx";
+import {createNotificationRead, getNotifications} from "@/core/services/notificationService.ts";
+import {getErrorMessage} from "@/core/utils/errorHandler.ts";
+import {toast} from "@/components/ui/use-toast.ts";
 
 interface NotificationPanelProps {
-    notifications: Notification[];
     onClose: () => void;
-    onMarkAsRead: (id: number) => void;
-    onMarkAllAsRead: () => void;
+    setUnreadCount: React.Dispatch<React.SetStateAction<number>>;
 }
 
-export const NotificationPopover = ({notifications, onClose, onMarkAsRead, onMarkAllAsRead}: NotificationPanelProps) => {
+export const NotificationPopover = ({onClose, setUnreadCount}: NotificationPanelProps) => {
     const [dialogOpen, setDialogOpen] = useState(false);
     const [selectedNotification, setSelectedNotification] = useState<Notification>();
+    const [notifications, setNotifications] = useState<Notification[]>([]);
+    const [errorMessage, setErrorMessage] = useState<string | null>(null);
+
+    useEffect(() => {
+        fetchNotifications();
+    }, []);
+
+    const fetchNotifications = async () => {
+        try {
+            const response = await getNotifications({}, 1, 10);
+            setNotifications(response.contents);
+        } catch (error) {
+            const errorMessage = getErrorMessage(error as Error);
+            setErrorMessage(errorMessage);
+            toast({title: "Error", description: errorMessage, variant: "destructive",});        }
+    };
+
+    const markAsRead = async (id: number) => {
+        try {
+            await createNotificationRead([id]);
+
+            setNotifications((prevNotifications) =>
+                prevNotifications.map((n) => n.id === id ? {...n, status: NotificationStatus.READ} : n)
+            );
+            setUnreadCount((prev) => prev - 1);
+        } catch (error) {
+            const errorMessage = getErrorMessage(error as Error);
+            setErrorMessage(errorMessage);
+            toast({title: "Error", description: errorMessage, variant: "destructive",});
+        }
+    };
+
+    const markAllAsRead = async () => {
+        try {
+            const unreadIds = notifications
+                .filter((n) => n.status === NotificationStatus.SENT)
+                .map((n) => n.id);
+
+            if (unreadIds.length === 0) return;
+
+            await createNotificationRead(unreadIds);
+            setNotifications((prevNotifications) =>
+                prevNotifications.map((n) => ({...n, status: NotificationStatus.READ}))
+            );
+            setUnreadCount(0);
+        } catch (error) {
+            const errorMessage = getErrorMessage(error as Error);
+            setErrorMessage(errorMessage);
+            toast({title: "Error", description: errorMessage, variant: "destructive",});
+        }
+    };
 
     const handleNotificationClick = (notification: Notification) => {
-        onMarkAsRead(notification.id);
+        markAsRead(notification.id);
         setSelectedNotification(notification);
         setDialogOpen(true);
     };
@@ -40,7 +92,7 @@ export const NotificationPopover = ({notifications, onClose, onMarkAsRead, onMar
                                     <TooltipTrigger asChild>
                                         <button
                                             className="hover:bg-gray-200 rounded-full p-1.5 transition-colors text-blue-600"
-                                            onClick={onMarkAllAsRead}
+                                            onClick={markAllAsRead}
                                             aria-label="Mark all as read"
                                         >
                                             <Check className="h-4 w-4"/>
